Change time slider to select index instead
The time selection slider in the dashboard was updated. Previously, it was directly using the timestamps from the 'times' list. However, this version was prone to issues such as inaccurate slider positions due to millisecond variations in the timestamps. Instead, the slider now selects an index position in the sorted 'times' list, granting more accurate selection of element, which in turn results in more precise filtering of data.