Add color_mapper function to map values to colors
A new function named 'color_mapper' has been added to the main.py file in the pitak_dash directory. This function assigns a color ('green', 'yellow', 'orange', or 'red') to a given value. The coloration is helpful in categorizing and visually distinguishing varied data sets on the Folium map based on the assigned thresholds.