Modify time selection to range and enhance map zoom level
Extended the time selection to allow users select a range (start and end time) rather than a single time period. Added a warning if end time selected is lesser than start time. This provides users with more flexibility in selecting the time range for which they want to view the data. Also, increased the initial zoom level of the folium map to render a more detailed view.