Refined time input and added clickable markers on heatmap
Modified the time input widget to provide minute precision for start and end times and filtered the data based on this input. Added clickable markers on the heatmap displaying important data such as temperature, humidity, and risk probability for corresponding location. These changes will provide higher precision in data selection and more information for the user on the heatmap.