Replace time_input widget with selectbox in pitak_dash
The time input widget for start and end time selection is replaced by selectbox dropdown. The purpose of this transition is to use the unique times and create a dropdown time range. This change simplifies the UI and makes it more user friendly. The resulting selected time is used to filter data and compute average risk. Map view is left undisturbed, ensuring it functions as before.