Add smoke and fire image display in split layout
A new feature has been added to display smoke and fire images side by side in a single row. This was achieved by using the beta_columns function to divide the layout into two columns. The time of image display was determined by the current date and time, using the strftime function. This enhancement allows for more efficient visual comparison between the two types of images.