Refactor code for readability and fixing date format
The string concatenation used to build the tileurl was formatted across several lines for better readability. In addition, the date format was corrected in 'current_time'. Previously, it only showed the time, now it displays the date and time for better context in "Smoke Detection".  This is especially beneficial for tracking and logging purposes.